requirements.txt
#scikit dependencies
six
numpy
scipy

pymongo
matplotlib
tornado
scikit-image
#Necesario libjpeg-dev para poder leer jpeg
pillow
